Don't recalculate meshnode normals unnecessarily

+++ src/client/mesh.cpp
@@ -272,9 +272,9 @@ void rotateMeshYZby(scene::IMesh *mesh, f64 degrees)
 	rotateMesh<&v3f::Y, &v3f::Z>(mesh, degrees);
 }
 
-void rotateMeshBy6dFacedir(scene::IMesh *mesh, int facedir)
+void rotateMeshBy6dFacedir(scene::IMesh *mesh, u8 facedir)
 {
-	int axisdir = facedir >> 2;
+	u8 axisdir = facedir >> 2;
 	facedir &= 0x03;
 	switch (facedir) {
 		case 1: rotateMeshXZby(mesh, -90); break;
